requirements.txt
streamlit
openai
firecrawl-py==1.6.3
python-dotenv>=1.0.0
